requirements.txt
mlflow==2.4.2
torch==2.0.0
torch-geometric==2.3.1
torchmetrics==1.0.0
torchvision==0.15.1
pydantic==1.10.7
SQLAlchemy==2.0.1
tensorboard==2.12.3
pandas==1.5.3
numpy==1.23.5
aiohttp==3.8.4
python-binance==1.0.17
line-bot-sdk==3.1.0
